デバッグ実行対応
